Inline 'in_head.js' and remove from spec helper
We can't reference this file in production. Doing so causes Sprockets to
fail because it tries to interpret the asset but does not have the
a Javascript runtime to do so successfully.

Delete mas_analytics - this is supposed to be a unit test and should
not depend on global objects. We need to start again with this test.