Add toolTrackingId to GA dataLayer.
This was on the old website but never migrated over. We're changing it
to use dataLayer.push() for GTM while we're at it. The Budget Planner
and soon hopefully the Health Check, Redundancy Tool and Quiz engines
are meant to set the content for this to be a string that identifies the
tool along with a unique id for user's data. For example 'BP123456'.