Better Contento connection info in error
When trying to run Front End from development, there's a few things that
you need to have set up to get a connection to Contento. This is written
up in the README, but if someone's missed something it's not that easy
to tell them what they've missed.

This is an attempt to add a bit more context for debugging this error.