Remove Gaffe gem and adding custom exceptions
Using the gaffe implementation, the controller is called directly
and was not working render errors for engines. The url_for becomes
lost with the params in two different controllers and actions - the
controller requested and the error controller.
So add the normal routes to the exception app solves the issue and
mapping the error status by hand will solve the issue.