Added default parse response behaviour to all Resources
Now the resource module parses the response using
the attribute builder class. This parser will be used by default
by all entity classes that includes Resource module

Don't assume that response body objects are arrays
Without this, the codes assumes that the response will return
an array of hashes, it makes some specs on resource module to fail,
since they rely on simple strings, so we decide to commit separated.